+++ 52n-sos-coding-sensorML-v101/src/main/java/org/n52/sos/encode/GmlEncoderv311.java
@@ -499,7 +499,7 @@ public class GmlEncoderv311 implements IEncoder<XmlObject, Object> {
             measureType.setUom("");
         }
         if (!quantityValue.getValue().equals(Double.NaN)) {
-            measureType.setDoubleValue(quantityValue.getValue());
+            measureType.setDoubleValue(quantityValue.getValue().doubleValue());
         } else {
             measureType.setNil();
         }
